[CMake] Bugfix: do not look for random versions of LLVM.
find_package(LLVM) will alter LLVM_DIR if no candidate is found. As a
consequence, if the first call does not find the correct version of
LLVM, subsequent calls will not be able to consider the path manually
supplied by the user. We can fix this by restoring the original value of
LLVM_DIR if this happens.

LLVM's minor version number is effectively unused right now, so it's
fine to only check for one particular major.minor combination and move
to the default case when we don't find a valid version immediately.